+++ src/azure/service_bus/event_hub/mod.rs
@@ -1,6 +1,8 @@
 use azure::core::errors::{AzureError, UnexpectedHTTPResult};
 
 use hyper;
+use hyper::net::HttpsConnector;
+use hyper_native_tls::NativeTlsClient;
 use hyper::header::{Headers, ContentLength};
 use hyper::status::StatusCode;
 
@@ -8,9 +10,12 @@ use chrono;
 use time::Duration;
 
 use std::ops::Add;
-use rustc_serialize::base64::{STANDARD, ToBase64};
+use base64;
 
-use url::percent_encoding::{utf8_percent_encode, HTTP_VALUE_ENCODE_SET, FORM_URLENCODED_ENCODE_SET};
+use url::percent_encoding::utf8_percent_encode;
+use url::form_urlencoded::Serializer;
+
+use hyper::header::parsing::HTTP_VALUE;
 
 use crypto::sha2::Sha256;
 use crypto::hmac::Hmac;
@@ -28,7 +33,7 @@ fn send_event(namespace: &str,
               event_hub: &str,
               policy_name: &str,
               hmac: &mut Hmac<Sha256>,
-              event_body: (&mut Read, u64),
+              event_body: &mut (&mut Read, u64),
               duration: Duration)
               -> Result<(), AzureError> {
 
@@ -42,14 +47,17 @@ fn send_event(namespace: &str,
     // create content
 
     // generate sas signature based on key name, key value, url and duration.
-    let sas = generate_signature(&policy_name, hmac, &url.to_string(), duration);
+    let sas = generate_signature(policy_name, hmac, &url.to_string(), duration);
     debug!("sas == {}", sas);
 
     // add required headers (in this case just the Authorization and Content-Length).
-    let client = hyper::client::Client::new();
+    let ssl = NativeTlsClient::new().unwrap();
+    let connector = HttpsConnector::new(ssl);
+    let client = hyper::client::Client::with_connector(connector);
     let mut headers = Headers::new();
     headers.set(Authorization(sas));
     headers.set(ContentLength(event_body.1));
+    debug!("headers == {:?}", headers);
 
     let body = hyper::client::Body::SizedBody(event_body.0, event_body.1);
 
@@ -80,19 +88,27 @@ fn generate_signature(policy_name: &str,
     let expiry = chrono::UTC::now().add(ttl).timestamp();
     debug!("expiry == {:?}", expiry);
 
-    let url_encoded = utf8_percent_encode(url, HTTP_VALUE_ENCODE_SET);
-    debug!("url_encoded == {:?}", url_encoded);
+    let url_encoded = utf8_percent_encode(url, HTTP_VALUE);
+    //debug!("url_encoded == {:?}", url_encoded);
 
     let str_to_sign = format!("{}\n{}", url_encoded, expiry);
     debug!("str_to_sign == {:?}", str_to_sign);
 
     hmac.reset();
     hmac.input(str_to_sign.as_bytes());
-    let sig = hmac.result().code().to_base64(STANDARD);
-    let sig = utf8_percent_encode(&sig, FORM_URLENCODED_ENCODE_SET);
+    let sig = {
+        let sig = base64::encode(hmac.result().code());
+        debug!("sig == {}", sig);
+        let mut ser = Serializer::new(String::new());
+        ser.append_pair("sig", &sig);
+        let sig = ser.finish();
+        debug!("sig == {}", sig);
+        sig
+    };
+
     debug!("sig == {:?}", sig);
 
-    format!("SharedAccessSignature sr={}&sig={}&se={}&skn={}",
+    format!("SharedAccessSignature sr={}&{}&se={}&skn={}",
             &url_encoded,
             sig,
             expiry,